Move to stack for building
This library has been updated to use stack,
with the most recent resolver that has a version
of Happstack server.

This version uses base-4.12, where Semigroup is a
superclass of Monoid, so a few instances had to be
added and modified accordingly.

Addiitionally, there was no default implementation of
inputBool, so that was added so we could have a build
without warnings.