[ch16765] Remove Happstack dependency
The Happstack dependency and the exported Happstack.Aeson
module have been removed from this verson.

The Happstack.Aeson code will either live in a separate package
or be added to the two applications that depended on it.

Having it in this library forced consumers of this validation
package to also take a dependency on Happstack, even if
Happstack was not the web framework of choice