Aeson 2 and GHC 9.2 compatibility
GHC 9.2 added a new warning: '-Wnoncanonical-monoid-instances'.  Since
the warning is enabled by '-Wall', and we have '-Werror' on, compilation
fails. This commit resolves the issue by fixing the errors.

Aeson 2 refactors Object such that it isn't necessarily a HashMap.  To
adapt to these changes, we use CPP to conditionally enable code that
works with Aeson 2. Aeson 1 compatibility is retained.